+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added `reset_and_start` to `measure`
 * added `str` and `repr` to `AtomicNumber`
+* `ReprableObject` -> `ReprableMixin` is now far more useful